// packages/gameSchema/src/types.ts
export type IntegerOrInput =
  | number
  | { readonly $input: string }
  | { readonly $input: string; readonly ifTrue: number; readonly ifFalse: number }

export type InputValue = number | string | boolean

export interface PoolDefinition {
  readonly sides: IntegerOrInput
  readonly quantity?: IntegerOrInput
}

export interface PoolCondition {
  readonly pool?: 'preModify' | 'postModify'
  readonly countWhere: {
    readonly operator: '>' | '>=' | '<' | '<=' | '='
    readonly value: IntegerOrInput
  }
  readonly atLeast?: IntegerOrInput
  readonly atLeastRatio?: number
}

export interface TableRange {
  readonly result: string
  readonly min?: number
  readonly max?: number
  readonly exact?: number
  readonly poolCondition?: PoolCondition
}

export interface TableDefinition {
  readonly ranges: readonly TableRange[]
}

export interface Ref {
  readonly $ref: string
}

export type RefOrTableDefinition = Ref | TableDefinition
// ...
export interface DegreeOfSuccessOperation {
  readonly criticalSuccess?: number
  readonly success?: number
  readonly failure?: number
  readonly criticalFailure?: number
}

export type OutcomeOperation =
  | { readonly ranges: readonly TableRange[] }
  | { readonly degreeOfSuccess: DegreeOfSuccessOperation }
  | { readonly tableLookup: RefOrTableDefinition }
// ...
export interface RandSumSpec {
  readonly $schema: string
  readonly name: string
  readonly shortcode: string
  readonly version?: string
  readonly game_url: string
  readonly srd_url?: string
  readonly pools?: Readonly<Record<string, PoolDefinition>>
  readonly tables?: Readonly<Record<string, TableDefinition>>
  readonly outcomes?: Readonly<Record<string, OutcomeOperation>>
  readonly [key: string]: unknown
}